requirements.txt
# Old versions (deprecated)
# langchain==0.1.0
# wikipedia==1.4.0
# langchain-community==0.0.20
# langchain-openai==0.0.8
# langchain-anthropic==0.0.2
# python-dotenv==1.0.0
# pydantic==2.5.0
# duckduckgo-search==5.1.0

# Current dependencies
langchain
langchain-core
langchain-community
langchain-classic
langchain-google-genai
wikipedia==1.4.0
python-dotenv
pydantic
ddgs